+++ lib/routes/slim.js
@@ -1,3 +1,3 @@
-var slim = require('../transforms/libxml/slim')
+var slim = require('../transforms/slim')
 
 module.exports = require('./restbase-req-handler.js')(slim)
