+++ lib/transforms/slim.js
@@ -19,7 +19,7 @@ module.exports = (html) => {
   return doc.toString()
 }
 
-function removeAll(nodes) {
+function removeAll (nodes) {
   var i
 
   for (i = 0; i < nodes.length; ++i) {
